- Refactor codebase to use new structure - Rename `proving_system` to `proof_system`
