+++ src/proof_system/linearisation_poly.rs
@@ -1,5 +1,5 @@
 use crate::fft::{EvaluationDomain, Polynomial};
-use crate::proving_system::PreProcessedCircuit;
+use crate::proof_system::PreProcessedCircuit;
 use bls12_381::Scalar;
 #[cfg(feature = "serde")]
 use serde::{de::Visitor, ser::SerializeStruct, Deserialize, Deserializer, Serialize, Serializer};
