+++ src/proof_system/mod.rs
@@ -3,8 +3,8 @@
 pub(crate) mod linearisation_poly;
 pub(crate) mod preprocessed_circuit;
 pub(crate) mod proof;
-pub(crate) mod prover;
 pub(crate) mod quotient_poly;
 pub(crate) mod widget;
 
 pub use preprocessed_circuit::PreProcessedCircuit;
+pub use proof::Proof;
