+++ src/proof_system/quotient_poly.rs
@@ -1,7 +1,7 @@
 use crate::fft::{EvaluationDomain, Polynomial};
 /// This quotient polynomial can only be used for the standard composer
 /// Each composer will need to implement their own method for computing the quotient polynomial
-use crate::proving_system::PreProcessedCircuit;
+use crate::proof_system::PreProcessedCircuit;
 use bls12_381::Scalar;
 use rayon::prelude::*;
