+++ src/proof_system/widget/permutation.rs
@@ -5,7 +5,7 @@ use super::PreProcessedPolynomial;
 use crate::commitment_scheme::kzg10::Commitment;
 use crate::fft::{EvaluationDomain, Evaluations, Polynomial};
 use crate::permutation::constants::{K1, K2, K3};
-use crate::proving_system::linearisation_poly::ProofEvaluations;
+use crate::proof_system::linearisation_poly::ProofEvaluations;
 use bls12_381::{G1Affine, Scalar};
 use rayon::prelude::*;
 #[cfg(feature = "serde")]
